pyproject.toml
[project]
name = "embedded-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "psutil>=7.2.2",
    "pymicro-wakeword>=1.0.0",
    "rich>=14.3.3",
    "sounddevice>=0.5.5",
    "uvicorn>=0.41.0",
    "pycaw>=2.0.1",
    "comtypes>=1.4.9",
    "pillow>=11.1.0",
    "screen-brightness-control>=0.23.0",
    "rapidfuzz>=3.12.1",
    "vosk>=0.3.45",
]
